apply to hooks
