+++ python_modules/libraries/dagster-msteams/dagster_msteams/card.py
@@ -1,8 +1,7 @@
 from collections.abc import Mapping
-from typing import TYPE_CHECKING, Optional
+from typing import Optional
 
-if TYPE_CHECKING:
-    from dagster_msteams.client import Link
+from dagster_msteams.utils import Link
 
 
 class Card:
@@ -27,7 +26,7 @@ class Card:
         content_type = "application/vnd.microsoft.card.hero"
         return {"contentType": content_type, "content": content}
 
-    def add_attachment(self, text_message: str, link: Optional["Link"] = None):
+    def add_attachment(self, text_message: str, link: Optional[Link] = None):
         if link:
             text_message += f" <a href='{link.url}'>{link.text}</a>"
         hero_card_attachment = self._create_attachment(text_message)
