+++ python_modules/libraries/dagster-msteams/dagster_msteams/utils.py
@@ -1,0 +1,6 @@
+from typing import NamedTuple
+
+
+class Link(NamedTuple):
+    text: str
+    url: str
